requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiormq==6.4.2
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.9.1.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.1
bleach==5.0.1
blis==0.7.9
boto3==1.28.22
botocore==1.31.22
CacheControl==0.12.11
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.35.0
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.0.4
confluent-kafka==1.9.2
contourpy==1.0.5
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
dask==2022.10.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.4
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-embed-video==1.4.9
django-filter==23.2
django-import-export==3.3.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-taggit==4.0.0
django-templated-mail==1.1.1
djangoql==0.17.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
dnspython==1.16.0
docopt==0.6.2
drf-yasg==1.21.7
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
fastjsonschema==2.16.2
fbmessenger==6.0.0
filelock==3.12.0
fire==0.5.0
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h5py==3.7.0
httptools==0.5.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
inflection==0.5.1
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jax==0.4.14
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
locket==1.0.0
magic-filter==1.0.9
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mistune==2.0.4
ml-dtypes==0.2.0
msgpack==1.0.5
multidict==5.2.0
murmurhash==1.0.9
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pamqp==3.2.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.28
protobuf==4.23.3
psutil==5.9.3
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
Pygments==2.13.0
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.18.1
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
questionary==1.10.0
randomname==0.1.5
rasa-sdk==3.6.1
redis==4.5.5
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.14.0
shiboken6==6.5.1.1
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.21.3
smart-open==6.3.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.48
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.5.1
structlog==23.1.0
structlog-sentry==2.0.3
tablib==3.5.0
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.1
terminado==0.17.0
terminaltables==3.1.10
thinc==8.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
transformers==4.26.0
twilio==7.14.2
typeguard==3.0.2
typer==0.7.0
typing-utils==0.1.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
